requirements.txt
pydantic
psycopg2-binary
python-dotenv
pytest
fastapi
uvicorn[standard]
httpx
watchdog
websockets
